Started using pregenerated common emuns, fixed a mysql field length issue with titlelists

+++ src/servers/Server_Zone/Network/PacketWrappers/ActorSpawnPacket.h
@@ -29,7 +29,7 @@ private:
    {
       // TODO: temporary gm rank
       m_data.gmRank = 0xff;
-      m_data.classJob = pPlayer->getClass();
+      m_data.classJob = static_cast< uint8_t >( pPlayer->getClass() );
       m_data.status = static_cast< uint8_t >( pPlayer->getStatus() );
       m_data.hPCurr = pPlayer->getHp();
       m_data.mPCurr = pPlayer->getMp();
